Start to consume os-client-config
There are two libraries - os-client-config and shade - that already
exist for the purposes of what these utility libs are doing.

os-client-config's job is to collect client config information, whether
via environment variables, command line arguments or clouds.yaml files
and construct relevant objects from that. For instance, it will very
happily make novaclient.client.Client objects for user - and it'll
handle all of the arguments possible.

shade's job is to contain business logic. So a good amount of the API
calls in NovaClient are probably just things that need to live in shade.
That's a few more steps than this is, but since shade is also driven by
os-client-config, moving to os-client-config as a first step seems
fairly sane.